requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic==2.10.3
python-dotenv==1.2.2
psycopg2-binary==2.9.9
